requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
attrs==21.2.0
    # via pytest
click==8.0.1
    # via code-annotations
code-annotations==1.1.2
    # via -r requirements/test.in
coverage==5.5
    # via pytest-cov
ddt==1.4.2
    # via -r requirements/test.in
django-model-utils==4.1.1
    # via -r requirements/base.txt
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-model-utils
iniconfig==1.1.1
    # via pytest
jinja2==3.0.1
    # via code-annotations
markupsafe==2.0.1
    # via jinja2
packaging==20.9
    # via pytest
pbr==5.6.0
    # via stevedore
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pyparsing==2.4.7
    # via packaging
pytest-cov==2.12.1
    # via -r requirements/test.in
pytest-django==4.4.0
    # via -r requirements/test.in
pytest==6.2.4
    # via
    #   pytest-cov
    #   pytest-django
python-slugify==5.0.2
    # via code-annotations
pytz==2021.1
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==5.4.1
    # via code-annotations
sqlparse==0.4.1
    # via
    #   -r requirements/base.txt
    #   django
stevedore==3.3.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
